LayoutTests/inspector/uncaught-dom1-exception.html

<html>
<head>
<script src="evaluate-in-frontend.js"></script>
<script>

function doit()
{
    // Since the exception will kill execution, make a short timeout.
    setTimeout(function() {
        alert("Exception timeout exceeded.")
        if (window.layoutTestController) {
            layoutTestController.closeWebInspector();
            layoutTestController.notifyDone();
        }
    }, 500);

    var p = document.createTextNode();
    p.textContent = "foo";
    p.splitText(10000);
}

</script>
</head>

<body onload="onload()">
<p>
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
</p>

<div id="frontend-script" style="display:none">
</div>

<div id="output">
</div>

</body>
</html>
