LayoutTests/inspector/uncaught-dom1-exception-expected.txt

CONSOLE MESSAGE: line 19: INDEX_SIZE_ERR: DOM Exception 1: Index or size was negative, or greater than the allowed value.
ALERT: Exception timeout exceeded.
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
